Introduce Invoice data structure.
The idea is to create a chain of rules to calculate the total.

Each rule captures an Invoice does it job and returns an Invoice ( so
they can be chained )

This would allow to apply multiples rules for the invoice generation
process.

Next commit I'll try to calculate total using the invoice structure.